# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TAU Service API",
    "description": "TAU Service Template API documentation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get current user",
        "description": "Returns the currently authenticated user's profile",
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/equipment/import": {
      "post": {
        "tags": [
          "Equipment"
        ],
        "summary": "Import equipment records from CSV",
        "description": "Bulk-creates Equipment Catalog entries from a CSV payload. Returns the imported count and any per-row errors. Requires admin, editor or rto role. rto is admitted at the route layer and narrowed per record in the service: the catalog holds both satcom and radio equipment, separated by the terminal_type column rather than by route, so a scoped radio writer whose file carries a satcom row gets 403 EQUIPMENT_RADIO_SCOPE_ONLY rather than introducing records the drawer would refuse.",
        "operationId": "importEquipment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CsvImportRequest"
              }
            }
          }
        },
        "responses": {
          "207": {
            "description": "Import processed - some or all rows were written, with any per-row errors listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsvImportResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body, empty CSV, or every row failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - admin, editor or rto role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/export/equipment": {
      "get": {
        "tags": [
          "Equipment"
        ],
        "summary": "Export equipment records as CSV",
        "description": "Streams a filtered CSV of Equipment Catalog entries. Read-only; open to any authenticated user.",
        "operationId": "exportEquipment",
        "parameters": [
          {
            "name": "terminal_type",
            "in": "query",
            "description": "Restrict to one terminal type - satcom or radio (empty = all)",
            "schema": {
              "type": "string",
              "enum": [
                "satcom",
                "radio"
              ]
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Free-text catalog search (empty = all)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "columns",
            "in": "query",
            "description": "Comma-separated column subset (empty = all, canonical order)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV file returned",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/equipment": {
      "get": {
        "tags": [
          "Equipment"
        ],
        "summary": "List equipment",
        "operationId": "listEquipment",
        "description": "Returns the Equipment Catalog. Open to any authenticated user.",
        "responses": {
          "200": {
            "description": "Equipment retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListEquipmentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Equipment"
        ],
        "summary": "Create an equipment record",
        "operationId": "createEquipment",
        "description": "Requires admin, editor, or rto.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipmentResponse"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Equipment created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateEquipmentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. The caller is authenticated but lacks the role this route requires.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/equipment/{id}": {
      "get": {
        "tags": [
          "Equipment"
        ],
        "summary": "Get an equipment data sheet",
        "operationId": "getEquipment",
        "description": "Open to any authenticated user.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Equipment ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Equipment retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEquipmentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Equipment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Equipment"
        ],
        "summary": "Update an equipment record",
        "operationId": "updateEquipment",
        "description": "Requires admin, editor, or rto.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Equipment ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipmentResponse"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Equipment updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEquipmentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. The caller is authenticated but lacks the role this route requires.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Equipment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Equipment"
        ],
        "summary": "Delete an equipment record",
        "operationId": "deleteEquipment",
        "description": "Requires admin, editor, or rto.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Equipment ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Equipment deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. The caller is authenticated but lacks the role this route requires.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Equipment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token"
      }
    },
    "schemas": {
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "User's unique identifier"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address"
          },
          "name": {
            "type": "string",
            "description": "User's display name"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "User's assigned roles"
          },
          "preferences": {
            "$ref": "#/components/schemas/PreferencesResponse"
          },
          "last_login_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Last login timestamp"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Account creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "roles",
          "preferences",
          "created_at",
          "updated_at"
        ]
      },
      "PreferencesResponse": {
        "type": "object",
        "properties": {
          "theme": {
            "type": "string",
            "enum": [
              "light",
              "dark"
            ],
            "description": "User's preferred theme"
          }
        },
        "required": [
          "theme"
        ]
      },
      "GetUserResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "$ref": "#/components/schemas/UserResponse"
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Error code for programmatic handling"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ValidationError"
                },
                "description": "Validation error details (only present for validation errors)"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "description": "Field that failed validation"
          },
          "message": {
            "type": "string",
            "description": "Validation error message"
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
      "ImportRowError": {
        "type": "object",
        "properties": {
          "row": {
            "type": "integer",
            "description": "1-based CSV row number"
          },
          "name": {
            "type": "string",
            "description": "Name value from the offending row"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Validation errors for this row"
          }
        },
        "required": [
          "row",
          "name",
          "errors"
        ]
      },
...
      "CsvImportRequest": {
        "type": "object",
        "description": "The CSV import body shared by every domain that imports through csvhttp: equipment, waveforms, services, transports and nets. Terminals and kits predate it and keep their own request schemas.",
        "properties": {
          "csv": {
            "type": "string",
            "description": "Raw CSV content to import"
          }
        },
        "required": [
          "csv"
        ]
      },
      "CsvImportResponse": {
        "type": "object",
        "description": "The import result shared by every domain that imports through csvhttp. 207 whenever anything was written, including a completely clean import; 400 only when every row failed.",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "imported": {
                "type": "integer",
                "description": "Number of rows written"
              },
              "errors": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ImportRowError"
                }
              },
              "message": {
                "type": "string",
                "description": "Human-readable import summary"
              }
            }
          }
        }
      },
...
      "EquipmentResponse": {
        "type": "object",
        "description": "A catalog data sheet for a SATCOM or radio terminal.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nomenclature": {
            "type": "string"
          },
          "nickname": {
            "type": "string",
            "nullable": true
          },
          "one_liner": {
            "type": "string",
            "nullable": true
          },
          "doc_number": {
            "type": "string",
            "nullable": true
          },
          "photo_url": {
            "type": "string",
            "nullable": true,
            "description": "Read through the authenticated photo endpoint below, never as a raw blob URL."
          },
          "make": {
            "type": "string",
            "nullable": true
          },
          "terminal_type": {
            "type": "string"
          },
          "operational_mode": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "data": {
            "type": "object",
            "additionalProperties": true,
            "description": "Free-form specification blob: frequencies, gain and EIRP, power, SWAP, features and the compatibility matrix. Held as JSONB because the fields differ by terminal type."
          },
          "created_by": {
            "type": "string"
          },
          "updated_by": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "nomenclature",
          "terminal_type",
          "operational_mode",
          "data"
        ]
      },
      "ListEquipmentResponse": {
        "type": "object",
        "properties": {
          "equipment": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EquipmentResponse"
            }
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "equipment",
          "total"
        ]
      },
      "GetEquipmentResponse": {
        "type": "object",
        "properties": {
          "equipment": {
            "$ref": "#/components/schemas/EquipmentResponse"
          }
        },
        "required": [
          "equipment"
        ]
      },
      "CreateEquipmentResponse": {
        "type": "object",
        "properties": {
          "equipment": {
            "$ref": "#/components/schemas/EquipmentResponse"
          }
        },
        "required": [
          "equipment"
        ]
      },
...
    }
  },
...
}